requirements.txt
python-nmap>=0.7.1
pandas>=1.5.3
lxml>=4.9.1
